Move table parsing code to own file
